requirements.txt
xlrd3>=1.1.0
mysql-connector-python>=8.0.28
PyQt5>=5.15.2
